Behold main-configen av resttemplate som før, legg kun til testconfig som ikke bruker proxy mot microsoft
