+++ src/main/kotlin/no/nav/familie/ef/mottak/config/ApplicationConfig.kt
@@ -2,6 +2,9 @@ package no.nav.familie.ef.mottak.config
 
 import com.fasterxml.jackson.module.kotlin.KotlinModule
 import no.nav.familie.ef.mottak.api.filter.RequestTimeFilter
+import no.nav.familie.http.interceptor.BearerTokenClientInterceptor
+import no.nav.familie.http.interceptor.ConsumerIdClientInterceptor
+import no.nav.familie.http.interceptor.MdcValuesPropagatingClientInterceptor
 import no.nav.familie.log.filter.LogFilter
 import no.nav.security.token.support.client.spring.oauth2.EnableOAuth2Client
 import no.nav.security.token.support.spring.api.EnableJwtTokenValidation
@@ -9,10 +12,13 @@ import org.slf4j.LoggerFactory
 import org.springframework.boot.SpringBootConfiguration
 import org.springframework.boot.autoconfigure.domain.EntityScan
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan
+import org.springframework.boot.web.client.RestTemplateBuilder
 import org.springframework.boot.web.servlet.FilterRegistrationBean
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.ComponentScan
+import org.springframework.context.annotation.Import
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories
+import org.springframework.web.client.RestOperations
 import springfox.documentation.swagger2.annotations.EnableSwagger2
 
 @SpringBootConfiguration
@@ -25,9 +31,30 @@ import springfox.documentation.swagger2.annotations.EnableSwagger2
 @EnableOAuth2Client(cacheEnabled = true)
 @EnableSwagger2
 @EnableJwtTokenValidation(ignore = ["org.springframework", "springfox.documentation.swagger.web.ApiResourceController"])
+@Import(BearerTokenClientInterceptor::class,
+        MdcValuesPropagatingClientInterceptor::class,
+        ConsumerIdClientInterceptor::class)
 class ApplicationConfig {
 
     private val logger = LoggerFactory.getLogger(ApplicationConfig::class.java)
+
+
+    @Bean("restTemplateAzure")
+    fun restTemplateAzure(restTemplateBuilder: RestTemplateBuilder,
+                          mdcInterceptor: MdcValuesPropagatingClientInterceptor,
+                          bearerTokenClientInterceptor: BearerTokenClientInterceptor,
+                          consumerIdClientInterceptor: ConsumerIdClientInterceptor): RestOperations {
+        return restTemplateBuilder.interceptors(mdcInterceptor,
+                                                bearerTokenClientInterceptor,
+                                                consumerIdClientInterceptor).build()
+    }
+
+    @Bean("restTemplateUnsecured")
+    fun restTemplate(restTemplateBuilder: RestTemplateBuilder,
+                     mdcInterceptor: MdcValuesPropagatingClientInterceptor,
+                     consumerIdClientInterceptor: ConsumerIdClientInterceptor): RestOperations {
+        return restTemplateBuilder.interceptors(mdcInterceptor, consumerIdClientInterceptor).build()
+    }
 
     @Bean
     fun kotlinModule(): KotlinModule = KotlinModule()
