+++ src/test/kotlin/no/nav/familie/ef/mottak/config/RestTemplateTestConfig.kt
@@ -8,6 +8,7 @@ import org.springframework.boot.web.client.RestTemplateBuilder
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Import
+import org.springframework.context.annotation.Primary
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter
 import org.springframework.web.client.RestOperations
 import java.time.Duration
@@ -20,6 +21,7 @@ import java.time.Duration
 class RestTemplateTestConfig {
 
     @Bean("restTemplateAzure")
+    @Primary
     fun restTemplateJwtBearer(consumerIdClientInterceptor: ConsumerIdClientInterceptor,
                               bearerTokenClientInterceptor: BearerTokenClientInterceptor): RestOperations {
         return RestTemplateBuilder()
@@ -32,6 +34,7 @@ class RestTemplateTestConfig {
     }
 
     @Bean
+    @Primary
     fun restTemplateBuilderUtenProxy(consumerIdClientInterceptor: ConsumerIdClientInterceptor,
                                     mdcValuesPropagatingClientInterceptor: MdcValuesPropagatingClientInterceptor
     ): RestTemplateBuilder {
